+++ helix-dap/src/lib.rs
@@ -1,5 +1,4 @@
 mod client;
-mod test;
 mod transport;
 mod types;
